Add time and working scoped limits (#1854)
* Add time and working time scoped limit support.

* Add regression test for working time check when tearing down sandbox.

* update docs and changelog

* Update CHANGELOG.md

---------

+++ src/inspect_ai/log/_transcript.py
@@ -111,7 +111,7 @@ class SampleLimitEvent(BaseEvent):
     message: str
     """A message associated with this limit"""
 
-    limit: int | None = Field(default=None)
+    limit: float | None = Field(default=None)
     """The limit value (if any)"""
